requirements.txt
discord.py
typer
python-dotenv
aiosqlite